+++ backend/appengine/routes/encontros/encontro.py
@@ -9,13 +9,16 @@ from routes import encontros
 from tekton import router
 from tekton.gae.middleware.redirect import RedirectResponse
 from gaecookie.decorator import no_csrf
-from gaepermission.decorator import permissions
-from permission_app.model import COORDENADOR, ADMIN, CATEQUISTA
+from gaepermission.decorator import login_required
+from permission_app.model import COORDENADOR, validate_permission, CATEQUIZANDO
 
 
-@permissions(ADMIN, COORDENADOR)
+@login_required
 @no_csrf
-def index(id=0):
+def index(_logged_user, id=0):
+    access_denid = validate_permission(CATEQUIZANDO, _logged_user)
+    if access_denid:
+        return access_denid
     context = {}
     try:
         key_id = int(id)
@@ -32,9 +35,12 @@ def index(id=0):
     return TemplateResponse(context, template_path='/encontros/encontro.html')
 
 
-@permissions(ADMIN, COORDENADOR)
+@login_required
 @no_csrf
-def save(**encontro_properties):
+def save(_logged_user, **encontro_properties):
+    access_denid = validate_permission(COORDENADOR, _logged_user)
+    if access_denid:
+        return access_denid
     cmd = encontro_facade.save_encontro_cmd(**encontro_properties)
     try:
         cmd()
@@ -46,9 +52,12 @@ def save(**encontro_properties):
     return RedirectResponse(router.to_path(encontros))
 
 
-@permissions(ADMIN, CATEQUISTA)
+@login_required
 @no_csrf
-def edit(**encontro_properties):
+def edit(_logged_user, **encontro_properties):
+    access_denid = validate_permission(COORDENADOR, _logged_user)
+    if access_denid:
+        return access_denid
     obj_id = encontro_properties.pop("key_id", None)
     cmd = encontro_facade.update_encontro_cmd(obj_id, **encontro_properties)
     try:
@@ -61,9 +70,12 @@ def edit(**encontro_properties):
     return RedirectResponse(router.to_path(encontros))
 
 
-@permissions(ADMIN, COORDENADOR)
+@login_required
 @no_csrf
-def delete(obj_id=0):
+def delete(_logged_user, obj_id=0):
+    access_denid = validate_permission(COORDENADOR, _logged_user)
+    if access_denid:
+        return access_denid
     cmd = encontro_facade.delete_encontro_cmd(obj_id)
     try:
         cmd()
