+++ backend/appengine/routes/noticias/noticia.py
@@ -3,8 +3,9 @@ from __future__ import absolute_import, unicode_literals
 from time import sleep
 from config.template_middleware import TemplateResponse
 from gaebusiness.business import CommandExecutionException
-from gaepermission.decorator import login_not_required
+from gaepermission.decorator import login_not_required, login_required
 from noticia_app.noticia_model import Noticia
+from permission_app.model import validate_permission, CATEQUIZANDO, COORDENADOR
 from tekton import router
 from gaecookie.decorator import no_csrf
 from noticia_app import noticia_facade
@@ -12,9 +13,12 @@ from routes import noticias
 from tekton.gae.middleware.redirect import RedirectResponse
 
 
-@login_not_required
+@login_required
 @no_csrf
-def index(id=0):
+def index(_logged_user, id=0):
+    access_denid = validate_permission(CATEQUIZANDO, _logged_user)
+    if access_denid:
+        return access_denid
     context = {}
     try:
         key_id = int(id)
@@ -30,9 +34,12 @@ def index(id=0):
     context["nav_active"] = 'noticias'
     return TemplateResponse(context, template_path='/noticias/noticia.html')
 
-@login_not_required
+@login_required
 @no_csrf
-def save(**noticia_properties):
+def save(_logged_user, **noticia_properties):
+    access_denid = validate_permission(COORDENADOR, _logged_user)
+    if access_denid:
+        return access_denid
     cmd = noticia_facade.save_noticia_cmd(**noticia_properties)
     try:
         cmd()
@@ -43,9 +50,12 @@ def save(**noticia_properties):
     return RedirectResponse(router.to_path(noticias))
 
 
-@login_not_required
+@login_required
 @no_csrf
-def edit(**noticia_properties):
+def edit(_logged_user, **noticia_properties):
+    access_denid = validate_permission(COORDENADOR, _logged_user)
+    if access_denid:
+        return access_denid
     obj_id = noticia_properties.pop("key_id", None)
     cmd = noticia_facade.update_noticia_cmd(obj_id, **noticia_properties)
     try:
@@ -57,9 +67,12 @@ def edit(**noticia_properties):
     return RedirectResponse(router.to_path(noticias))
 
 
-@login_not_required
+@login_required
 @no_csrf
-def delete(obj_id=0):
+def delete(_logged_user, obj_id=0):
+    access_denid = validate_permission(COORDENADOR, _logged_user)
+    if access_denid:
+        return access_denid
     cmd = noticia_facade.delete_noticia_cmd(obj_id)
     try:
         cmd()
